requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
confection==0.0.2
cymem==2.0.6
datasets==2.5.1
decorator==5.1.1
dill==0.3.5.1
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.1
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.28
happytransformer==2.4.1
huggingface-hub==0.10.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
langcodes==3.3.0
MarkupSafe==2.1.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
mysql-connector-python==8.0.30
numpy==1.23.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pathy==0.6.2
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
preshed==3.0.7
protobuf==3.20.1
pyarrow==9.0.0
pydantic==1.9.2
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
responses==0.18.0
rich==12.6.0
semver==2.13.0
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.0/en_core_web_md-3.4.0.tar.gz
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
streamlit==1.13.0
thinc==8.1.2
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
tornado==6.2
tqdm==4.64.1
transformers==4.22.2
typer==0.4.2
typing-extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wasabi==0.10.1
watchdog==2.1.9
Werkzeug==2.2.2
xxhash==3.0.0
yarl==1.8.1
zipp==3.9.0
